refactor fixed hour tariff ui
